refactor: improve code structure
